Adding unit tests

+++ src/main/java/com/rentalCars/rentalCars/domain/cars/dto/CarDTO.java
@@ -32,7 +32,7 @@ public record CarDTO(
          @NotBlank
          String rentedBy,
          @NotNull
-         BigDecimal price,
+         Double price,
          @NotBlank
          String airbag,
          @NotBlank
